id re-created with UID to prevent password reset being used repeatedly Fragment logic moved
